# tp-final/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants/{tenantSlug}/reservas": {
      "post": {
        "operationId": "ReservasController_create",
        "parameters": [
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos invalidos o fecha muy lejana.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro o el tratamiento no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Sin cupo, fuera de agenda, horario pasado o turno duplicado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Reserva un turno, con o sin cuenta. Con token de admin, el centro lo carga a nombre de una clienta.",
        "tags": [
          "Reservas"
        ]
      },
      "get": {
        "operationId": "ReservasController_findAll",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pendiente",
                "cancelada",
                "confirmada",
                "ausente"
              ]
            }
          },
          {
            "name": "servicioId",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "name": "desde",
            "required": false,
            "in": "query",
            "description": "Desde esta fecha, inclusive.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "required": false,
            "in": "query",
            "description": "Hasta esta fecha, inclusive.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Cuantos elementos devolver como maximo en esta pagina.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "El nextCursor que devolvio la pagina anterior. Es opaco: mandalo tal cual.",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Una pagina de resultados.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CursorPageDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ReservaDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro, o de una cuenta que no es de administracion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "La agenda del centro, paginada por cursor. Solo la administradora.",
        "tags": [
          "Reservas"
        ]
      }
    },
    "/api/v1/tenants/{tenantSlug}/reservas/{reservaId}": {
      "get": {
        "operationId": "ReservasController_findOne",
        "parameters": [
          {
            "name": "reservaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaDto"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La reserva no existe, o no es de la clienta de la sesion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Una reserva. La ve su duenia o la administracion.",
        "tags": [
          "Reservas"
        ]
      },
      "patch": {
        "operationId": "ReservasController_update",
        "parameters": [
          {
            "name": "reservaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservaDto"
                }
              }
            }
          },
          "400": {
            "description": "Estado y fecha juntos, fecha sin hora, o reembolsar fuera de lugar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Token de otro centro, o una clienta pidiendo algo que es del centro.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "La reserva no existe, o no es de la clienta de la sesion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "Transicion invalida, fuera de plazo para reprogramar, sin cupo en el horario nuevo o ausente antes de hora.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancela, reprograma, confirma o marca ausente un turno.",
        "tags": [
          "Reservas"
        ]
      }
    },
    "/api/v1/tenants/{tenantSlug}/reservas/{reservaId}/estado": {
      "get": {
        "operationId": "ReservasController_estado",
        "parameters": [
          {
            "name": "reservaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoReservaSoloDto"
                }
              }
            }
          },
          "404": {
            "description": "La reserva no existe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "El estado de una reserva, sin datos personales. Para la pagina de vuelta de un pago.",
        "tags": [
          "Reservas"
        ]
      }
    },
    "/api/v1/tenants/{tenantSlug}/clientes/me/reservas": {
      "get": {
        "operationId": "MisReservasController_findAll",
        "parameters": [
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pendiente",
                "cancelada",
                "confirmada",
                "ausente"
              ]
            }
          },
          {
            "name": "servicioId",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 40,
              "type": "string"
            }
          },
          {
            "name": "desde",
            "required": false,
            "in": "query",
            "description": "Desde esta fecha, inclusive.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "required": false,
            "in": "query",
            "description": "Hasta esta fecha, inclusive.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Cuantos elementos devolver como maximo en esta pagina.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "El nextCursor que devolvio la pagina anterior. Es opaco: mandalo tal cual.",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Una pagina de resultados.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CursorPageDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ReservaDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro, o no es de una clienta.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Los turnos de la clienta, tambien los que reservo sin cuenta con el mismo email.",
        "tags": [
          "MisReservas"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de turnos",
    "description": "SaaS multi-tenant de turnos para centros de estetica. Cada centro es un tenant y viaja en el path. Los montos son enteros en centavos. Las horas son hora de pared del centro, en formato HH:mm, y las fechas YYYY-MM-DD.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Identificador de maquina del error. Es el que conviene usar para ramificar.",
            "enum": [
              "validation_error",
              "invalid_cursor",
              "invalid_credentials",
              "unauthenticated",
              "invalid_token",
              "wrong_tenant",
              "tenant_not_found",
              "servicio_not_found",
              "reserva_not_found",
              "servicio_name_taken",
              "ventanas_superpuestas",
              "slot_full",
              "outside_business_hours",
              "past_date",
              "too_far_ahead",
              "invalid_transition",
              "high_contention",
              "forbidden_role",
              "not_found",
              "payload_too_large",
              "too_many_requests",
              "internal_error",
              "http_error",
              "invalid_code",
              "too_many_codes",
              "cliente_not_found",
              "online_payment_unavailable",
              "payment_provider_unavailable",
              "reschedule_not_allowed",
              "too_early_for_no_show",
              "invalid_state",
              "mp_account_change_blocked",
              "mercadopago_not_configured",
              "monthly_limit_reached",
              "plan_limit_reached",
              "mercadopago_not_connected",
              "invalid_signature",
              "slug_taken"
            ],
            "type": "string",
            "example": "slot_full"
          },
          "message": {
            "type": "string",
            "description": "Texto en espanol, listo para mostrarle a una persona.",
            "example": "Ese horario ya no tiene cupo. Elegi otro."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CursorPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pasalo como `cursor` para pedir la pagina siguiente. null = era la ultima.",
            "example": null
          }
        },
        "required": [
          "data",
          "nextCursor"
        ]
      },
...
      "CreateReservaDto": {
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "description": "Formato YYYY-MM-DD.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2026-10-05"
          },
          "hora": {
            "type": "string",
            "description": "Formato HH:mm, de la grilla del centro.",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "example": "09:45"
          },
          "metodoPago": {
            "enum": [
              "efectivo",
              "mercadopago"
            ],
            "type": "string",
            "example": "efectivo"
          },
          "servicioId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "clienteNombre": {
            "type": "string",
            "description": "Un solo campo, como el formulario del prototipo. Obligatorio sin sesion de clienta.",
            "minLength": 2,
            "maxLength": 120
          },
          "clienteEmail": {
            "type": "string",
            "description": "La identidad de la clienta: la reserva queda asociada a este email aunque no tenga\ncuenta. Obligatorio sin sesion de clienta; con sesion, sale de la cuenta.",
            "format": "email"
          },
          "clienteTelefono": {
            "type": "string",
            "description": "Obligatorio sin sesion de clienta.",
            "minLength": 6,
            "maxLength": 30
          },
          "notas": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "fecha",
          "hora",
          "metodoPago",
          "servicioId"
        ]
      },
      "CobroDto": {
        "type": "object",
        "properties": {
          "montoCentavos": {
            "type": "number",
            "description": "La sena, o el precio total si se paga con Mercado Pago, en centavos.",
            "example": 540000
          },
          "pagadoCentavos": {
            "type": "number",
            "description": "Lo que entro por Mercado Pago, aprobado.",
            "example": 540000
          },
          "reembolsadoCentavos": {
            "type": "number",
            "description": "Lo que se devolvio.",
            "example": 0
          },
          "venceAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "Hasta cuando se puede pagar; despues la reserva vence. null si ya no esta pendiente."
          },
          "checkoutUrl": {
            "type": "string",
            "nullable": true,
            "description": "El link de Mercado Pago al que se redirige para pagar. null si ya no esta pendiente."
          }
        },
        "required": [
          "montoCentavos",
          "pagadoCentavos",
          "reembolsadoCentavos",
          "venceAt",
          "checkoutUrl"
        ]
      },
      "ReservaDto": {
        "type": "object",
        "properties": {
          "fecha": {
            "type": "string",
            "example": "2026-10-05"
          },
          "horaInicio": {
            "type": "string",
            "example": "09:45"
          },
          "horaFin": {
            "type": "string",
            "example": "10:30"
          },
          "estado": {
            "type": "string",
            "description": "`pendiente` (esperando el pago), `confirmada`, `cancelada` o `ausente`.",
            "example": "confirmada"
          },
          "canceladaPor": {
            "type": "string",
            "nullable": true,
            "description": "`clienta`, `centro` o `sistema` (vencio sin pagarse). null si no esta cancelada.",
            "example": null
          },
          "metodoPago": {
            "type": "string",
            "description": "`efectivo` o `mercadopago`.",
            "example": "efectivo"
          },
          "senaCentavos": {
            "type": "number",
            "description": "Sena congelada al momento de reservar, en centavos.",
            "example": 540000
          },
          "precioCentavos": {
            "type": "number",
            "description": "Precio congelado al momento de reservar, en centavos.",
            "example": 1800000
          },
          "cobro": {
            "nullable": true,
            "description": "El cobro online. null si no hay nada que pagar por Mercado Pago.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CobroDto"
              }
            ]
          },
          "reprogramacionHorasAntes": {
            "type": "number",
            "nullable": true,
            "description": "La politica que la clienta acepto al reservar. null: no puede reprogramar sola.",
            "example": 24
          },
          "cancelacionHorasAntes": {
            "type": "number",
            "example": 24
          },
          "id": {
            "type": "string"
          },
          "puedeReprogramar": {
            "type": "boolean",
            "description": "Si la clienta puede reprogramarla sola ahora. Lo calcula el servidor con la hora del\ncentro, para que el front no tenga que hacer aritmetica de zonas horarias."
          },
          "puedeCancelarConReembolso": {
            "type": "boolean",
            "description": "Si cancelarla ahora devuelve lo pagado. Fuera de plazo se pierde todo lo pagado."
          },
          "clienteNombre": {
            "type": "string"
          },
          "clienteEmail": {
            "type": "string"
          },
          "clienteTelefono": {
            "type": "string"
          },
          "notas": {
            "type": "string",
            "nullable": true
          },
          "servicioId": {
            "type": "string"
          },
          "clienteId": {
            "type": "string",
            "description": "La clienta duenia del turno. Existe aunque haya reservado sin cuenta."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "fecha",
          "horaInicio",
          "horaFin",
          "estado",
          "canceladaPor",
          "metodoPago",
          "senaCentavos",
          "precioCentavos",
          "cobro",
          "reprogramacionHorasAntes",
          "cancelacionHorasAntes",
          "id",
          "puedeReprogramar",
          "puedeCancelarConReembolso",
          "clienteNombre",
          "clienteEmail",
          "clienteTelefono",
          "notas",
          "servicioId",
          "clienteId",
          "createdAt"
        ]
      },
      "EstadoReservaSoloDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "description": "`pendiente`, `confirmada`, `cancelada` o `ausente`.",
            "example": "confirmada"
          }
        },
        "required": [
          "estado"
        ]
      },
      "UpdateReservaDto": {
        "type": "object",
        "properties": {
          "estado": {
            "description": "Transiciones validas: pendiente -> confirmada | cancelada, confirmada -> cancelada |\nausente. `cancelada` y `ausente` son terminales.",
            "enum": [
              "confirmada",
              "cancelada",
              "ausente"
            ],
            "type": "string",
            "example": "cancelada"
          },
          "fecha": {
            "type": "string",
            "description": "Fecha nueva, para reprogramar. Va con `hora`.",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2026-10-12"
          },
          "hora": {
            "type": "string",
            "description": "Hora nueva, de la grilla del centro. Va con `fecha`.",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "example": "15:00"
          },
          "reembolsar": {
            "type": "boolean",
            "description": "Solo el centro, solo al cancelar: si se devuelve lo pagado. Obligatorio si hubo pago.\nPara la clienta lo decide la politica del servicio."
          }
        }
      },
...
    }
  }
}